move edax midgame logic to separate file
