+++ src/bot/edax/mod.rs
@@ -1,3 +1,4 @@
 pub mod bot;
 pub mod eval;
+pub mod midgame;
 pub mod weights;
